Flush out settings handling

+++ applications/system-service/application.cpp
@@ -36,7 +36,7 @@ Application::launchWithArgs(const QStringList& args) {
 }
 void
 Application::launchNoSecurityCheck(const QStringList& args) {
-  if (m_process->processId()) {
+  if (m_process == nullptr || m_process->processId()) {
     resumeNoSecurityCheck();
   } else {
     if (sharedSettings.applicationUsage()) {
@@ -161,8 +161,8 @@ Application::pause(bool startIfNone) {
 void
 Application::pauseNoSecurityCheck(bool startIfNone) {
   if (
-    !m_process->processId() || stateNoSecurityCheck() == Paused ||
-    type() == Background
+    m_process == nullptr || !m_process->processId() ||
+    stateNoSecurityCheck() == Paused || type() == Background
   ) {
     return;
   }
@@ -193,8 +193,8 @@ Application::pauseNoSecurityCheck(bool startIfNone) {
 void
 Application::interruptApplication() {
   if (
-    !m_process->processId() || stateNoSecurityCheck() == Paused ||
-    type() == Background
+    m_process == nullptr || !m_process->processId() ||
+    stateNoSecurityCheck() == Paused || type() == Background
   ) {
     return;
   }
@@ -281,6 +281,9 @@ Application::waitForResume() {
 
 QString
 Application::id() {
+  if (m_process == nullptr) {
+    return QString();
+  }
   return QString("connection/%1").arg(m_process->processId());
 }
 
@@ -303,7 +306,8 @@ Application::resume() {
 void
 Application::resumeNoSecurityCheck() {
   if (
-    !m_process->processId() || stateNoSecurityCheck() == InForeground ||
+    m_process == nullptr || !m_process->processId() ||
+    stateNoSecurityCheck() == InForeground ||
     (type() == Background && stateNoSecurityCheck() == InBackground)
   ) {
     O_DEBUG("Can't Resume" << path() << "Already running!");
@@ -334,7 +338,8 @@ Application::resumeNoSecurityCheck() {
 void
 Application::uninterruptApplication() {
   if (
-    !m_process->processId() || stateNoSecurityCheck() == InForeground ||
+    m_process == nullptr || !m_process->processId() ||
+    stateNoSecurityCheck() == InForeground ||
     (type() == Background && stateNoSecurityCheck() == InBackground)
   ) {
     return;
@@ -505,9 +510,10 @@ Application::stopNoSecurityCheck() {
     }
   );
 }
+
 void
 Application::signal(int signal) {
-  if (m_process->processId()) {
+  if (m_process != nullptr && m_process->processId()) {
     kill(-m_process->processId(), signal);
   }
 }
@@ -536,6 +542,9 @@ Application::unregisterNoSecurityCheck() {
 
 int
 Application::stateNoSecurityCheck() {
+  if (m_process == nullptr) {
+    return Inactive;
+  }
   switch (m_process->state()) {
     case QProcess::Starting:
     case QProcess::Running: {
@@ -702,6 +711,7 @@ Application::finished(int exitCode) {
   getCompositorDBus()->lower(
     QString("connection/%1").arg(m_process->processId())
   );
+  m_process = nullptr;
   if (
     exitCode != EXIT_SUCCESS && exitCode != SIGINT && exitCode != SIGTERM &&
     exitCode != SIGKILL && exitCode != 128 + EXIT_SUCCESS &&
@@ -734,6 +744,9 @@ Application::finished(int exitCode) {
 
 void
 Application::readyReadStandardError() {
+  if (m_process == nullptr) {
+    return;
+  }
   QString error = m_process->readAllStandardError();
   if (p_stderr != nullptr) {
     *p_stderr << error.toStdString().c_str();
@@ -757,6 +770,9 @@ Application::readyReadStandardError() {
 
 void
 Application::readyReadStandardOutput() {
+  if (m_process == nullptr) {
+    return;
+  }
   QString output = m_process->readAllStandardOutput();
   if (p_stdout != nullptr) {
     *p_stdout << output.toStdString().c_str();
@@ -863,7 +879,8 @@ void
 Application::delayUpTo(int milliseconds) {
   timer.invalidate();
   timer.start();
-  while (timer.isValid() && !timer.hasExpired(milliseconds)) {
+  while (m_process != nullptr && timer.isValid() &&
+         !timer.hasExpired(milliseconds)) {
     QCoreApplication::processEvents(QEventLoop::AllEvents, 100);
   }
 }
@@ -909,7 +926,9 @@ Application::updateEnvironment() {
   for (auto key : environment().keys()) {
     env.insert(key, environment().value(key, "").toString());
   }
-  m_process->setEnvironment(env.toStringList());
+  if (m_process != nullptr) {
+    m_process->setEnvironment(env.toStringList());
+  }
 }
 
 void
@@ -930,7 +949,7 @@ Application::startSpan(std::string operation, std::string description) {
 
 void
 Application::waitForFinished() {
-  if (m_process->processId()) {
+  if (m_process != nullptr && m_process->processId()) {
     m_process->waitForFinished();
   }
 }
@@ -1031,6 +1050,9 @@ Application::name() {
 }
 int
 Application::processId() {
+  if (m_process == nullptr) {
+    return 0;
+  }
   return m_process->processId();
 }
 QStringList
